moved LinearCacheManager, added caching to ScipyKrylov
