+++ openmdao/visualization/n2_viewer/tests/sellar_final_values.json
@@ -29,7 +29,9 @@
         "err_on_non_converge": false,
         "assemble_jac": false,
         "solver": "gmres",
-        "restart": 20
+        "restart": 20,
+        "use_cache": true,
+        "max_cache_entries": 3
     },
     "component_type": null,
     "subsystem_type": "group",
@@ -131,7 +133,9 @@
                 "err_on_non_converge": false,
                 "assemble_jac": false,
                 "solver": "gmres",
-                "restart": 20
+                "restart": 20,
+                "use_cache": true,
+                "max_cache_entries": 3
             },
             "component_type": null,
             "subsystem_type": "group",
@@ -155,7 +159,9 @@
                         "err_on_non_converge": false,
                         "assemble_jac": false,
                         "solver": "gmres",
-                        "restart": 20
+                        "restart": 20,
+                        "use_cache": true,
+                        "max_cache_entries": 3
                     },
                     "component_type": null,
                     "subsystem_type": "group",
@@ -539,16 +545,16 @@
                     "implicit": false,
                     "units": "None",
                     "val": [
-                        28.588308164980585
+                        28.58830816498059
                     ],
                     "val_min_indices": [
                         0
                     ],
-                    "val_min": 28.588308164980585,
+                    "val_min": 28.58830816498059,
                     "val_max_indices": [
                         0
                     ],
-                    "val_max": 28.588308164980585
+                    "val_max": 28.58830816498059
                 }
             ],
             "options": {
