+++ openmdao/visualization/n2_viewer/tests/sellar_no_values_run.json
@@ -29,7 +29,9 @@
         "err_on_non_converge": false,
         "assemble_jac": false,
         "solver": "gmres",
-        "restart": 20
+        "restart": 20,
+        "use_cache": true,
+        "max_cache_entries": 3
     },
     "component_type": null,
     "subsystem_type": "group",
@@ -108,7 +110,9 @@
                 "err_on_non_converge": false,
                 "assemble_jac": false,
                 "solver": "gmres",
-                "restart": 20
+                "restart": 20,
+                "use_cache": true,
+                "max_cache_entries": 3
             },
             "component_type": null,
             "subsystem_type": "group",
@@ -132,7 +136,9 @@
                         "err_on_non_converge": false,
                         "assemble_jac": false,
                         "solver": "gmres",
-                        "restart": 20
+                        "restart": 20,
+                        "use_cache": true,
+                        "max_cache_entries": 3
                     },
                     "component_type": null,
                     "subsystem_type": "group",
